Spread spawn positions to 54x54 tile area
Players now spawn up to 27 tiles away in both the x- and y-axis from the
force's spawn position. This should largely eliminate the ability for
enemies to spawn camp.

Also 1) moved the rocket silo to the center of the base, 2) replaced
"Factorio PvP" with "Silo Assault" in a few places, 4) reduced
num_alien_artifacts_gradual from 50 to 40, 4) reordered starting items
on toolbelt.